requirements.txt
osxphotos==0.74.1
pillow==10.4.0
ImageHash==4.3.1
Flask==3.0.3
